Finished Temporary Design of the admin login Page
